requirements.txt
# 실전 거래 봇 필수 라이브러리
ccxt>=4.0.0
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0  # 환경 변수 관리

# 백테스팅 (선택사항)
pandas_ta>=0.3.14b0

# GPU/CPU 최적화 라이브러리 (선택사항)
numba>=0.58.0
cupy-cuda12x>=12.0.0  # CUDA 12.x용 CuPy (CUDA 버전에 맞게 설치 필요)
scipy>=1.10.0

# 병렬 처리 (선택사항)
psutil>=5.9.0

# 머신러닝 라이브러리 (선택사항)
scikit-learn>=1.3.0
xgboost>=2.0.0  # GPU 지원 버전
lightgbm>=4.0.0  # GPU 지원 버전

# 시계열 분석 (선택사항)
statsmodels>=0.14.0

# 데이터 전처리 (선택사항)
imbalanced-learn>=0.11.0

